% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{strip_credentials}
\alias{strip_credentials}
\alias{creds}
\alias{get_credentials}
\title{Get or remove user authentication credentials}
\usage{
strip_credentials(urls)

get_credentials(urls)
}
\arguments{
\item{urls}{a URL, or vector of URLs}
}
\description{
authentication credentials appear before the domain
name and look like \emph{user:password}. Sometimes you want the removed,
or retrieved; \code{strip_credentials} and \code{get_credentials} do
precisely that
}
\examples{
# Remove credentials
strip_credentials("http://foo:bar@97.77.104.22:3128")

# Get credentials
get_credentials("http://foo:bar@97.77.104.22:3128")
}
