% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{domain}
\alias{domain}
\alias{domain<-}
\title{Get or set a URL's domain}
\usage{
domain(x)

domain(x) <- value
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{value}{a replacement value (or vector of replacement values)
for x's scheme.}
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://cran.r-project.org/submit.html"
domain(example_url)

#Set a component
domain(example_url) <- "en.wikipedia.org"
}
\seealso{
\code{\link{scheme}}, \code{\link{port}}, \code{\link{path}},
\code{\link{parameters}} and \code{\link{fragment}} for other accessors.
}
