% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tld.R
\name{domain_names}
\alias{domain_names}
\title{top-level domain names}
\format{two character vectors; one, \code{domain_names}, contains the lower-cased
TLDs. The other, \code{regex_ready_domain_names}, stores them in the format
"\\.domain$" to make them easily applicable to hostname parsing.}
\description{
a vector containing the accepted top-level domain names, retrieved from
\href{http://data.iana.org/TLD/tlds-alpha-by-domain.txt}{the IANA}.
}

