% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\name{scheme}
\alias{domain}
\alias{domain<-}
\alias{fragment}
\alias{fragment<-}
\alias{path}
\alias{path<-}
\alias{port}
\alias{port<-}
\alias{query}
\alias{query<-}
\alias{scheme}
\alias{scheme<-}
\title{Get or set component parts of URLs}
\usage{
scheme(x)

scheme(x) <- value

domain(x)

domain(x) <- value

port(x)

port(x) <- value

path(x)

path(x) <- value

query(x)

query(x) <- value

fragment(x)

fragment(x) <- value
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://cran.r-project.org/submit.html"
scheme(example_url)

#Set a component
scheme(example_url) <- "https"
}

