% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{parameters}
\alias{parameters}
\alias{parameters<-}
\title{Get or set a URL's parameters}
\usage{
parameters(x)

parameters(x) <- value
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{value}{a replacement value for x's parameters.}
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://en.wikipedia.org/wiki/Aaron_Halfaker?debug=true"
parameters(example_url)
#[1] "debug=true"

#Set a component
parameters(example_url) <- "debug=false"
}
\seealso{
\code{\link{scheme}}, \code{\link{domain}}, \code{\link{port}},
\code{\link{path}} and \code{\link{fragment}} for other accessors.
}

