% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\name{path}
\alias{path}
\alias{path<-}
\title{Get or set a URL's path}
\usage{
path(x)

path(x) <- value
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://cran.r-project.org:80/submit.html"
path(example_url)

#Set a component
path(example_url) <- "bin/windows/"
}

