% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\name{port}
\alias{port}
\alias{port<-}
\title{Get or set a URL's port}
\usage{
port(x)

port(x) <- value
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{value}{a replacement value for x's port.}
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://cran.r-project.org:80/submit.html"
port(example_url)

#Set a component
port(example_url) <- "12"
}
\seealso{
\code{\link{scheme}}, \code{\link{domain}}, \code{\link{path}},
\code{\link{parameters}} and \code{\link{fragment}} for other accessors.
}

