% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punycode.R
\name{punycode}
\alias{puny_decode}
\alias{puny_encode}
\alias{punycode}
\title{Puncycode Encoder/Decoder}
\usage{
puny_encode(s)

puny_decode(s)
}
\arguments{
\item{s}{A string or vector of strings.}
}
\value{
A string or vector of strings.
}
\description{
Encode/Decode a UTF-8 string to/from its punycode representation.  See references
for details.
}
\examples{
\dontrun{
library(urltools)
str <- "bücher"
(p <- topuny(str))
## [1] "bcher-kva"

unpuny(p)
## [1] "bücher"
}

}
\references{
Punycode RFC \url{http://www.faqs.org/rfcs/rfc3492.html}
Implementation \url{https://dxr.mozilla.org/mozilla-central/source/netwerk/dns/punycode.c}
}

