% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\name{scheme}
\alias{scheme}
\alias{scheme<-}
\title{Get or set a URL's scheme}
\usage{
scheme(x)

scheme(x) <- value
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://cran.r-project.org/submit.html"
scheme(example_url)

#Set a component
scheme(example_url) <- "https"
}

