% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/suffix.R
\docType{data}
\name{suffix_dataset}
\alias{suffix_dataset}
\title{Dataset of public suffixes}
\format{A vector of 7430 elements.}
\usage{
data(suffix_dataset)
}
\description{
This dataset contains a registry of public suffixes, as retrieved from
and defined by the \href{https://publicsuffix.org/}{public suffix list}. It is
sorted by how many periods(".") appear in the suffix, to optimise it for
\code{\link{suffix_extract}}.
}
\note{
Last updated 2015-05-06.
}
\seealso{
\code{\link{suffix_extract}} for extracting suffixes from domain names.
}
\keyword{datasets}

