% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/suffix.R
\name{suffix_refresh}
\alias{suffix_refresh}
\title{refresh the stored public suffix dataset}
\usage{
suffix_refresh()
}
\value{
TRUE if this worked; an error otherwise.
}
\description{
\code{urltools} comes with an inbuilt
dataset of public suffixes, \code{\link{suffix_dataset}}.
This is used in \code{\link{suffix_extract}} to identify the top-level domain
within a particular domain name.

While updates to the dataset will be included in each new package release,
there's going to be a gap between changes to TLDs and changes to the package.
Accordingly, the package also includes \code{\link{suffix_refresh}}, which refreshes
this dataset on the user end. This requires CURL (the library, not the
R package) on your system to function.
}
\examples{
\dontrun{
suffix_refresh()
}
}
\seealso{
\code{\link{suffix_extract}} to extract suffixes from domain names,
or \code{\link{suffix_dataset}} for the dataset itself.
}

