% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix.R
\docType{data}
\name{tld_dataset}
\alias{tld_dataset}
\title{Dataset of top-level domains (TLDs)}
\format{A vector of 1275 elements.}
\usage{
data(tld_dataset)
}
\description{
This dataset contains a registry of top-level domains, as retrieved from
and defined by the \href{http://data.iana.org/TLD/tlds-alpha-by-domain.txt}{IANA}.
}
\note{
Last updated 2016-07-20.
}
\seealso{
\code{\link{tld_extract}} for extracting TLDs from domain names,
and \code{\link{tld_refresh}} to get an updated version of this dataset.
}
\keyword{datasets}
