% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tld.R
\name{tld_refresh}
\alias{tld_refresh}
\title{refresh the stored Top Level Domain dataset}
\usage{
tld_refresh()
}
\value{
TRUE if this worked; an error otherwise.
}
\description{
\code{urltools} comes with an inbuilt
dataset of Top Level Domains (TLDs), \code{\link{tld_dataset}}.
This is used in \code{\link{tld_extract}} to identify the top-level domain
within a particular domain name.

While updates to the dataset will be included in each new package release,
there's going to be a gap between changes to TLDs and changes to the package.
Accordingly, the package also includes \code{\link{tld_refresh}}, which refreshes
this dataset on the user end. This requires CURL (the library, not the
R package) on your system to function.
}
\seealso{
\code{\link{tld_extract}} to extract TLDs from domain names,
or \code{\link{tld_dataset}} for the dataset itself.
#examples
\dontrun{
tld_refresh()
}
}

