% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tld.R
\name{update_domains}
\alias{update_domains}
\title{Update the locally-stored vector of top-level domains}
\usage{
update_domains()
}
\description{
\code{urltools} includes a vector of accepted top-level domains,
\code{\link{domain_names}}, which is used in \code{\link{domain_parse}}.
These should be updated with every new package release, but can
also be updated by the user by calling this function.

This functionality (and this update function) was extensively influenced by Jay Jacobs'
\href{https://github.com/jayjacobs/tldextract}{tldextract} package.
}
\examples{
\dontrun{
update_domains()
}
}
\seealso{
\code{\link{domain_parse}}, which uses this vector, and \code{\link{domain_names}}
for the dataset itself.
}

