% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{url_compose}
\alias{url_compose}
\title{Recompose Parsed URLs}
\usage{
url_compose(parsed_urls)
}
\arguments{
\item{parsed_urls}{a data.frame sourced from \code{\link{url_parse}}}
}
\description{
Sometimes you want to take a vector of URLs, parse them, perform
some operations and then rebuild them. \code{url_compose} takes a data.frame produced
by \code{\link{url_parse}} and rebuilds it into a vector of full URLs (or: URLs as full
as the vector initially thrown into url_parse).

This is currently a `beta` feature; please do report bugs if you find them.
}
\examples{
#Parse a URL and compose it
url <- "http://en.wikipedia.org"
url_compose(url_parse(url))

}
\seealso{
\code{\link{scheme}} and other accessors, which you may want to
run URLs through before composing them to modify individual values.
}
