% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urltools.R
\name{url_parameters}
\alias{url_parameters}
\title{get the values of a URL's parameters (deprecated)}
\usage{
url_parameters(urls, parameter_names)
}
\description{
the old name for \code{\link{param_get}}; deprecated
in favour of that function.
}

