% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_path.R
\name{url_path}
\alias{url_path}
\title{Construct URL Path}
\usage{
url_path(x, ...)
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{...}{character vectors}
}
\description{
Construct a URL from path components like \code{file.path}
}
\examples{

url_path("https://github.com", c('Ironholds', 'nfultz'), 'urltools')

}
