% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{urltools}
\alias{urltools}
\alias{urltools-package}
\title{Tools for handling URLs}
\description{
This package provides functions for URL encoding and decoding, parsing, and parameter extraction.
             The implementations are both fast and entirely vectorised It is intended to be useful for people
             making requests who need a fast, reliable encoding methods, and analysts dealing with data such
             as server-side logs.
}

