% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_host}
\alias{extract_host}
\title{extract a URL's hostname and provide it}
\usage{
extract_host(urls)
}
\arguments{
\item{urls}{a vector of URLs}
}
\value{
a vector of hostnames. In the event that a hostname cannot be identified, an
empty string will be returned.
}
\description{
similar to \code{\link{url_parse}} or \code{\link{extract_parameter}}, but
for hostnames only.
}

