% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{fragment}
\alias{fragment}
\alias{fragment<-}
\title{Get or set a URL's fragment}
\usage{
fragment(x)

fragment(x) <- value
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{value}{a replacement value (or vector of replacement values)
for x's fragment. If NULL, the fragment will be removed entirely.}
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://en.wikipedia.org/wiki/Aaron_Halfaker?debug=true#test"
fragment(example_url)

#Set a component
fragment(example_url) <- "production"

#Remove a component
fragment(example_url) <- NULL
}
\seealso{
\code{\link{scheme}}, \code{\link{domain}}, \code{\link{port}},
\code{\link{path}} and \code{\link{parameters}} for other accessors.
}
