% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{path}
\alias{path}
\alias{path<-}
\title{Get or set a URL's path}
\usage{
path(x)

path(x) <- value
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{value}{a replacement value (or vector of replacement values)
for x's path. If NULL, the path will be removed entirely.}
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
# Get the path
example_url <- "http://cran.r-project.org:80/submit.html"
path(example_url)

# Set the path
path(example_url) <- "bin/windows/"

# Remove the path
path(example_url) <- NULL
}
\seealso{
\code{\link{scheme}}, \code{\link{domain}}, \code{\link{port}},
\code{\link{parameters}} and \code{\link{fragment}} for other accessors.
}
