% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{port}
\alias{port}
\alias{port<-}
\title{Get or set a URL's port}
\usage{
port(x)

port(x) <- value
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{value}{a replacement value (or vector of replacement values)
for x's port. If NULL, the port will be entirely removed.}
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
# Get the port
example_url <- "http://cran.r-project.org:80/submit.html"
port(example_url)

# Set the port
port(example_url) <- "12"

# Remove the port
port(example_url) <- NULL
}
\seealso{
\code{\link{scheme}}, \code{\link{domain}}, \code{\link{path}},
\code{\link{parameters}} and \code{\link{fragment}} for other accessors.
}
