% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\name{query}
\alias{query}
\alias{query<-}
\title{Get or set a URL's query string}
\usage{
query(x)

query(x) <- value
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{value}{a replacement value for x's scheme.}
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://en.wikipedia.org/wiki/Aaron_Halfaker?debug=true"
query(example_url)

#Set a component
query(example_url) <- "debug=false"
}

