% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{scheme}
\alias{scheme}
\alias{scheme<-}
\title{Get or set a URL's scheme}
\usage{
scheme(x)

scheme(x) <- value
}
\arguments{
\item{x}{a URL, or vector of URLs}

\item{value}{a replacement value (or vector of replacement values)
for x's scheme.}
}
\description{
as in the lubridate package, individual components of a URL
can be both extracted or set using the relevant function call - see the
examples.
}
\examples{
#Get a component
example_url <- "http://cran.r-project.org/submit.html"
scheme(example_url)

#Set a component
scheme(example_url) <- "https"

# NA out the URL
scheme(example_url) <- NA_character_
}
\seealso{
\code{\link{domain}}, \code{\link{port}}, \code{\link{path}},
\code{\link{parameters}} and \code{\link{fragment}} for other accessors.
}
