% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/urltools.R
\docType{package}
\name{urltools}
\alias{urltools}
\alias{urltools-package}
\title{Tools for handling URLs}
\description{
This package provides functions for URL encoding and decoding,
parsing, and parameter extraction, designed to be both fast and
entirely vectorised. It is intended to be useful for people dealing with
web-related datasets, such as server-side logs.
}
\seealso{
the \href{http://cran.rstudio.com/web/packages/urltools/vignettes/urltools.html}{package vignette}.
}

